# Improve performance of creating Hash object