@Watson1978 2017-12-03
### Enviroment

* OS : Ubuntu 17.10
* CPU : Intel(R) Core(TM) i5-3210M CPU @ 2.50GHz
* Compiler : gcc version 7.2.0

### Before
```
$ ./miniruby -I../benchmark-ips/lib ~/tmp/bench/literal.rb
Warming up --------------------------------------
        Hash literal   137.004k i/100ms
Calculating -------------------------------------
        Hash literal      3.050M (± 0.5%) i/s -     15.344M in   5.030443s
```

### After
```
$ ./miniruby -I../benchmark-ips/lib ~/tmp/bench/literal.rb
Warming up --------------------------------------
        Hash literal   159.540k i/100ms
Calculating -------------------------------------
        Hash literal      3.882M (± 0.6%) i/s -     19.464M in   5.013813s
```