@Watson1978 2017-12-10
This is benchmark result with/without jemalloc.

	macOS  -> 38.5% up
	Ubuntu -> 29.0% up

## macOS
### Before
```
$ ./miniruby -I. -I../benchmark-ips/lib ~/tmp/bench/literal.rb
Warming up --------------------------------------
        Hash literal    65.472k i/100ms
Calculating -------------------------------------
        Hash literal      1.407M (± 1.6%) i/s -      7.071M in   5.026698s
```

### After
```
$ ./miniruby -I. -I../benchmark-ips/lib ~/tmp/bench/literal.rb
Warming up --------------------------------------
        Hash literal    78.338k i/100ms
Calculating -------------------------------------
        Hash literal      1.958M (± 1.9%) i/s -      9.792M in   5.002829s
```


## Ubuntu
### Before
```
$ ./miniruby -I. -I../benchmark-ips/lib ~/tmp/bench/literal.rb
Warming up --------------------------------------
        Hash literal   129.587k i/100ms
Calculating -------------------------------------
        Hash literal      3.055M (± 0.1%) i/s -     15.291M in   5.005374s
```

### After
```
$ ./miniruby -I. -I../benchmark-ips/lib ~/tmp/bench/literal.rb
Warming up --------------------------------------
        Hash literal   146.119k i/100ms
Calculating -------------------------------------
        Hash literal      3.918M (± 0.1%) i/s -     19.726M in   5.034367s
```